requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
dj-rest-auth==2.2.5
djangorestframework-simplejwt==4.7.1
django-rest-swagger==2.2.0
required==0.4.0
tzdata==2022.6
